requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
redis>=4.5
celery>=5.3
kombu>=5.3
pydantic>=2.5
python-decouple>=3.8
factory-boy>=3.2
Pillow>=10.0 